updating sample
