+++ backend/sample/package.json
@@ -14,6 +14,7 @@
   "dependencies": {
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
+    "mongodb": "^6.4.0",
     "mongoose": "^8.0.1",
     "nodemon": "^3.0.1"
   }
