+++ frontend/sample/index.html
@@ -14,7 +14,7 @@
       rel="stylesheet"
       href="//cdnjs.cloudflare.com/ajax/libs/leaflet.draw/1.0.3/leaflet.draw.css"
     />
-    <title>Vite + React + TS</title>
+    <title>SAMPLE UK HOUSE PRICE</title>
   </head>
   <body>
     <div id="root"></div>
